requirements.txt
asgiref==3.8.1
asyncio==3.4.3
channels==4.1.0
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
environs==11.0.0
marshmallow==3.21.3
numpy==2.0.0
packaging==24.1
pandas==2.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
websocket-client==1.8.0
